Updates to better handle large database sizes
After collecting more than a years worth of data, the process of reading all the records into memory was slowing down performance. This has been addressed by limiting the reads to 105 days at a time. When the user scrolls outside of this range, the read window is adjusted to include the range currently being displayed.