Fixing EGVRecord.testNum() for G4
This fixes the 'EGVRecord testNum #6' issue pointed out by jmimu. The new version of testNum() for G4 will always return 0. For G5 and G6, a real test number will be returned. We can't use testNum to identify calibration events on a G4 system. It may be possible to use separate Calibration records on G4, but that will be difficult to implement without access to a G4 device.