Avoid printing '.00' in Insulin amounts
This implements improvement 3 suggested by Testostheron1963-de in Issue #8, 'Three proposals to improve your software'. If the user specifies a fractional unit of Insulin (e.g. 10.05), it will be properly displayed, but if they specify a whole unit (3) or a tenth of a unit (3.1), trailing zeros will no longer be displayed. I've also bumped the release version number to 3.2.