Adding predictions for 1 hour in future
Adding code to find a polynomial which predicts future glucose values based on recent data points. These predicted values are added to the graph display. A two-degree polynomial is displayed in green. A one-degree polynomial is displayed in red. Also, the command : keymap.all_axes = '' has been removed, since it is now deprecated in matplotlib. The behavior it used to supply is now implemented by the addition of: ax.set_navigate(False)