Eliminate hang on shutdown
If the user closes the application by clicking on the window close button, it used to sometimes hang forever. I tracked this down to happening when the plotgraph() routine was running. To fix it, a new 'closeInProgress' variable has been added. When a close operation begins, this boolean is set to True. Any subsequent call to plotgraph() checks the variable, and returns immediately, if True. This eliminates the hang.