Adding -u option to swap units
For a Receiver using mg/dL units, adding '-u' to the command line will cause the glucose values to be displayed in mmol/L units. For a Receiver using mmol/L units, adding '-u' to the command line will cause the glucose values to be displayed in mg/dL units. Code for handling unit changes has been updated to stop it from crashing.