Removing serial port lock-up on Windows
On Unix-based operating systems, you can have two threads opening connections to the same port, but on Windows, they interfere with each other. To fix this, the threads have been updated to share the same serial port instance. This eliminates the problems running on Windows, and still works fine on Linux or MacOs.
In addition, the writes to the window title area seem to be working on Windows, now, so they are no longer skipped under that OS.
Also, several zorder values have been updated to fix some layering errors.