Fixing handling of -t-0:30 syntax
Setting a time offset of negative 30 minutes was failing because the minus sign got lost when converting -0 hours to an integer. Code has been updated to check for that minus sign, and apply it to minutes or seconds when the hours are set to zero.
Also updating the release version to 3.6 since many improvements have recently been added, and we want to encourage all users to update.