Fixing bug in Note display
After adding a new Note, the text would be duplicated at the next 5 minute update interval. The user would see a bold image of the text because the second copy is slightly offset. The user could drag one of the copies to clearly see the duplication. This problem was caused by having fractions of a second in the placement of the Note. Truncating the microseconds of the placement position removes this buggy behavior.