Update for newer releases of matplotlib, datetime.
dexctrack.py :
  In version 3.7.0 of matplotlib, the number of
positional arguments for Slider was switched from
6 to 5. Logic had to be added to check the version
and to handle both the older and newer forms of the
Slider API.

  In version 3.9.0 of matplotlib, the return value
of avxspan was switched from polygon to rectangle.
Logic had to be added to check the version, and
to handle both the old and new return forms.

  The method datetime.utcnow() is now deprecated,
and causes a warning to be output, so two calls
to it have been replaced with an alternative.

screensize.py :
  Support for the 'tkagg' backend has been added.

util.py :
  An unescaped backslash has been fixed to remove
a warning.

Resolves #17